ContextExtension вынесено в отдельный модуль
