+++ lib/app.dart
@@ -4,8 +4,8 @@ import 'package:places/domain/interactor/settings_interactor.dart';
 import 'package:places/ui/const/app_routes.dart';
 import 'package:places/ui/const/app_strings.dart';
 import 'package:places/ui/res/app_scope.dart';
+import 'package:places/ui/res/context_extension.dart';
 import 'package:places/ui/res/responsive.dart';
-import 'package:places/ui/res/theme_extension.dart';
 import 'package:places/ui/res/themes.dart';
 import 'package:places/ui/screen/splash_screen/splash_screen.dart';
 import 'package:provider/provider.dart';
