+++ lib/ui/screen/add_sight_screen/add_sight_screen_wm.dart
@@ -5,9 +5,9 @@ import 'package:places/domain/entity/sight.dart';
 import 'package:places/ui/const/app_routes.dart';
 import 'package:places/ui/const/app_strings.dart';
 import 'package:places/ui/res/app_scope.dart';
+import 'package:places/ui/res/context_extension.dart';
 import 'package:places/ui/res/logger.dart';
 import 'package:places/ui/res/scaffold_messenger_extension.dart';
-import 'package:places/ui/res/theme_extension.dart';
 import 'package:places/ui/screen/add_sight_screen/add_sight_screen.dart';
 import 'package:places/ui/screen/add_sight_screen/add_sight_screen_model.dart';
 import 'package:places/ui/widget/elementary/common_wm_mixin.dart';
