+++ lib/ui/screen/filters_screen/filters_screen_wm.dart
@@ -5,9 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:places/domain/entity/filter.dart';
 import 'package:places/domain/entity/sight.dart';
 import 'package:places/ui/res/app_scope.dart';
+import 'package:places/ui/res/context_extension.dart';
 import 'package:places/ui/res/logger.dart';
 import 'package:places/ui/res/scaffold_messenger_extension.dart';
-import 'package:places/ui/res/theme_extension.dart';
 import 'package:places/ui/screen/filters_screen/filters_screen.dart';
 import 'package:places/ui/screen/filters_screen/filters_screen_model.dart';
 import 'package:places/ui/screen/sight_list_screen/mixin/geo_permissions_wm_mixin.dart';
