+++ lib/ui/screen/settings_screen/settings_screen_wm.dart
@@ -2,7 +2,7 @@ import 'package:elementary/elementary.dart';
 import 'package:flutter/material.dart';
 import 'package:places/ui/const/app_routes.dart';
 import 'package:places/ui/res/app_scope.dart';
-import 'package:places/ui/res/theme_extension.dart';
+import 'package:places/ui/res/context_extension.dart';
 import 'package:places/ui/screen/settings_screen/settings_screen.dart';
 import 'package:places/ui/screen/settings_screen/settings_screen_model.dart';
 import 'package:places/ui/widget/elementary/common_wm_mixin.dart';
