+++ lib/ui/screen/sight_search_screen/sight_search_screen_wm.dart
@@ -3,9 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:places/domain/entity/sight.dart';
 import 'package:places/ui/const/app_routes.dart';
 import 'package:places/ui/res/app_scope.dart';
+import 'package:places/ui/res/context_extension.dart';
 import 'package:places/ui/res/logger.dart';
 import 'package:places/ui/res/scaffold_messenger_extension.dart';
-import 'package:places/ui/res/theme_extension.dart';
 import 'package:places/ui/screen/sight_search_screen/sight_search_screen.dart';
 import 'package:places/ui/screen/sight_search_screen/sight_search_screen_model.dart';
 import 'package:places/ui/widget/elementary/common_wm_mixin.dart';
